Feat: functionalize QuantTensor
